refactor: reservationroom -> reservationrooms
