--- src/main/java/com/beyond/easycheck/reservationroom/infrastructure/entity/ReservationStatus.java
@@ -1,7 +1,0 @@
-package com.beyond.easycheck.reservationroom.infrastructure.entity;
-
-public enum ReservationStatus {
-
-    RESERVATION,
-    CANCELED
-}
