+++ src/main/java/com/beyond/easycheck/reservationrooms/infrastructure/entity/PaymentStatus.java
@@ -1,0 +1,7 @@
+package com.beyond.easycheck.reservationrooms.infrastructure.entity;
+
+public enum PaymentStatus {
+
+    PAID,
+    UNPAID
+}
