+++ src/main/java/com/beyond/easycheck/reservationrooms/infrastructure/entity/ReservationStatus.java
@@ -1,0 +1,7 @@
+package com.beyond.easycheck.reservationrooms.infrastructure.entity;
+
+public enum ReservationStatus {
+
+    RESERVATION,
+    CANCELED
+}
