+++ src/main/java/com/beyond/easycheck/reservationrooms/infrastructure/repository/ReservationRoomRepository.java
@@ -1,6 +1,6 @@
-package com.beyond.easycheck.reservationroom.infrastructure.repository;
+package com.beyond.easycheck.reservationrooms.infrastructure.repository;
 
-import com.beyond.easycheck.reservationroom.infrastructure.entity.ReservationRoomEntity;
+import com.beyond.easycheck.reservationrooms.infrastructure.entity.ReservationRoomEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
