+++ src/main/java/com/beyond/easycheck/reservationrooms/ui/requestbody/ReservationRoomCreateRequest.java
@@ -1,7 +1,7 @@
-package com.beyond.easycheck.reservationroom.ui.requestbody;
+package com.beyond.easycheck.reservationrooms.ui.requestbody;
 
-import com.beyond.easycheck.reservationroom.infrastructure.entity.PaymentStatus;
-import com.beyond.easycheck.reservationroom.infrastructure.entity.ReservationStatus;
+import com.beyond.easycheck.reservationrooms.infrastructure.entity.PaymentStatus;
+import com.beyond.easycheck.reservationrooms.infrastructure.entity.ReservationStatus;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.validation.constraints.Min;
