+++ src/main/java/com/beyond/easycheck/reservationrooms/ui/requestbody/ReservationRoomUpdateRequest.java
@@ -1,6 +1,6 @@
-package com.beyond.easycheck.reservationroom.ui.requestbody;
+package com.beyond.easycheck.reservationrooms.ui.requestbody;
 
-import com.beyond.easycheck.reservationroom.infrastructure.entity.ReservationStatus;
+import com.beyond.easycheck.reservationrooms.infrastructure.entity.ReservationStatus;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import lombok.AllArgsConstructor;
