+++ src/main/java/com/beyond/easycheck/reservationservices/infrastructure/repository/ReservationServiceRepository.java
@@ -1,7 +1,7 @@
 package com.beyond.easycheck.reservationservices.infrastructure.repository;
 
 import com.beyond.easycheck.additionalservices.infrastructure.entity.AdditionalServiceEntity;
-import com.beyond.easycheck.reservationroom.infrastructure.entity.ReservationRoomEntity;
+import com.beyond.easycheck.reservationrooms.infrastructure.entity.ReservationRoomEntity;
 import com.beyond.easycheck.reservationservices.infrastructure.entity.ReservationServiceEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
